+++ tuxemon/npc.py
@@ -3,9 +3,9 @@
 from __future__ import annotations
 
 import logging
-from collections.abc import Iterable, Mapping, Sequence
+from collections.abc import Iterable, Sequence
 from math import hypot
-from typing import TYPE_CHECKING, Any, Optional, TypedDict
+from typing import TYPE_CHECKING, Optional
 
 from tuxemon.boxes import ItemBoxes, MonsterBoxes
 from tuxemon.db import DialogueProfile, Direction, NpcModel, db
@@ -28,6 +28,7 @@ from tuxemon.relationship import (
     decode_relationships,
     encode_relationships,
 )
+from tuxemon.save_state import NPCState
 from tuxemon.step_tracker import StepTrackerManager, decode_steps, encode_steps
 from tuxemon.teleporter import TeleportFaint
 from tuxemon.tools import vector2_to_tile_pos
@@ -42,30 +43,6 @@ if TYPE_CHECKING:
 
 
 logger = logging.getLogger(__name__)
-
-
-class NPCState(TypedDict, total=False):
-    current_map: str
-    facing: Direction
-    game_variables: dict[str, Any]
-    battles: Sequence[Mapping[str, Any]]
-    tuxepedia: Mapping[str, Any]
-    relationships: Mapping[str, Any]
-    money: Mapping[str, Any]
-    template: dict[str, Any]
-    missions: Sequence[Mapping[str, Any]]
-    items: Sequence[Mapping[str, Any]]
-    monsters: Sequence[Mapping[str, Any]]
-    player_name: str
-    player_steps: float
-    monster_boxes: dict[str, Sequence[Mapping[str, Any]]]
-    item_boxes: dict[str, Sequence[Mapping[str, Any]]]
-    tile_pos: tuple[int, int]
-    teleport_faint: tuple[str, int, int]
-    tracker: Mapping[str, Any]
-    step_tracker: Mapping[str, Any]
-    unlocked_letters: Mapping[str, Any]
-    evolution_registry: Mapping[str, Any]
 
 
 def tile_distance(tile0: Iterable[float], tile1: Iterable[float]) -> float:
@@ -167,7 +144,7 @@ class NPC(Entity[NPCState]):
 
         state: NPCState = {
             "current_map": session.client.get_map_name(),
-            "facing": self.facing,
+            "facing": self.facing.value,
             "game_variables": self._variables.get_player_state(),
             "battles": self.battle_handler.encode_battle(),
             "tuxepedia": encode_tuxepedia(self.tuxepedia),
