+++ tuxemon/session.py
@@ -5,9 +5,12 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING, Optional
 
+from tuxemon.save_state import NPCState, WorldSave
+
 if TYPE_CHECKING:
     from tuxemon.client import LocalPygameClient
     from tuxemon.player import Player
+    from tuxemon.save_state import SaveData
     from tuxemon.states.world_state import WorldState
 
 logger = logging.getLogger(__name__)
@@ -70,6 +73,36 @@ class Session:
             self._world = None
         if reset_player:
             self._player = None
+
+    def load_state(self, save_data: SaveData) -> None:
+        """
+        Loads the player, world, and other session-level states from a saved game dictionary.
+        """
+        if self._player is None or self._world is None:
+            raise ValueError(
+                "Player and World must be initialized before loading state."
+            )
+
+        self._player.set_state(self, save_data.get("npc_state", NPCState()))
+        self._world.set_state(self, save_data.get("world_state", WorldSave()))
+
+    def save_state(self, index: int, slot: int) -> SaveData:
+        """
+        Saves the player, world, and other session-level states to a dictionary.
+        """
+        if self._player is None or self._world is None:
+            raise ValueError(
+                "Player and World must be initialized before saving state."
+            )
+
+        # Local import to avoid circular dependency
+        from tuxemon import save
+
+        save_data = save.get_save_data(self)
+        save.save(save_data, index)
+        save.slot_number = slot
+
+        return save_data
 
 
 local_session = Session()
